Implement cache lookup of refresh token
We don't utilize the MSAL cache, so we need to obtain the refresh token
from the MSAL cache. We do this by generating the correct cache key and
looking it up manually.

Why don't we utilize the MSAL cache?
At this point, we don't utilize the MSAL cache because we want full
control over how tokens are cached and refreshed. Furthermore, the MSAL
token cache uses the user identity to cache tokens instead of a session
identity. In summary, implementing our own cache makes it easier to us
to control the behavior instead of having to work with the
particularities of the library.